# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BridgeAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workbench/analyze:
    post:
      operationId: analyzeWorkbenchTask
      tags: [workbench]
      summary: Analyse a training task
      description: >
        Accepts a platform task (instructions + content + optional rubric) and
        returns a recommended answer with rubric checklist, risk flags, confidence
        score, and review level. Every response must be reviewed by a human before
        manual submission. Automatic login, submission, or platform automation is
        not supported.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkbenchAnalyzeRequest"
      responses:
        "200":
          description: Analysis complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkbenchAnalyzeResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Request refused by safety policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workbench/refuse-check:
    post:
      operationId: workbenchRefuseCheck
      tags: [workbench]
      summary: Pre-flight safety check
      description: >
        Tests whether a given request text would be refused by the workbench
        safety policy without running a full analysis.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkbenchRefuseCheckRequest"
      responses:
        "200":
          description: Check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkbenchRefuseCheckResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    WorkbenchAnalyzeRequest:
      type: object
      properties:
        platform:
          type: string
          enum:
            - alignerr
            - outlier
            - dataannotation
            - toloka
            - remotasks
            - mindrift
            - other
        taskType:
          type: string
          enum:
            - grammar_cleanup
            - classification
            - sentiment_labeling
            - response_comparison
            - factuality_check
            - math_reasoning
            - coding
            - expert_domain
            - subjective_judgment
            - unknown
        instructions:
          type: string
        rubric:
          type: string
        taskContent:
          type: string
        answerOptions:
          type: array
          items:
            type: string
        userNotes:
          type: string
        budgetLimitUsd:
          type: number
        routingMode:
          type: string
          enum:
            - fast
            - balanced
            - quality
      required:
        - platform
        - instructions
        - taskContent
    WorkbenchAnalyzeResponse:
      type: object
      properties:
        taskId:
          type: string
        platform:
          type: string
        taskType:
          type: string
        recommendedAnswer:
          type: string
        confidence:
          type: number
        reasoningSummary:
          type: string
        riskFlags:
          type: array
          items:
            type: string
        rubricChecklist:
          type: array
          items:
            type: string
        reviewLevel:
          type: string
          enum:
            - quick_review
            - careful_review
            - human_only
        humanReviewRequired:
          type: boolean
        routingReceipt:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - taskId
        - platform
        - taskType
        - recommendedAnswer
        - confidence
        - reasoningSummary
        - riskFlags
        - rubricChecklist
        - reviewLevel
        - humanReviewRequired
    WorkbenchRefuseCheckRequest:
      type: object
      properties:
        requestText:
          type: string
      required:
        - requestText
    WorkbenchRefuseCheckResponse:
      type: object
      properties:
        allowed:
          type: boolean
        reason:
          type: string
          nullable: true
      required:
        - allowed
        - reason