# lib/api-zod/src/generated/types/workbenchAnalyzeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BridgeAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkbenchAnalyzeRequestPlatform } from "./workbenchAnalyzeRequestPlatform";
import type { WorkbenchAnalyzeRequestRoutingMode } from "./workbenchAnalyzeRequestRoutingMode";
import type { WorkbenchAnalyzeRequestTaskType } from "./workbenchAnalyzeRequestTaskType";

export interface WorkbenchAnalyzeRequest {
  platform: WorkbenchAnalyzeRequestPlatform;
  taskType?: WorkbenchAnalyzeRequestTaskType;
  instructions: string;
  rubric?: string;
  taskContent: string;
  answerOptions?: string[];
  userNotes?: string;
  budgetLimitUsd?: number;
  routingMode?: WorkbenchAnalyzeRequestRoutingMode;
}