# lib/api-zod/src/generated/types/workbenchAnalyzeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BridgeAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkbenchAnalyzeResponseReviewLevel } from "./workbenchAnalyzeResponseReviewLevel";
import type { WorkbenchAnalyzeResponseRoutingReceipt } from "./workbenchAnalyzeResponseRoutingReceipt";

export interface WorkbenchAnalyzeResponse {
  taskId: string;
  platform: string;
  taskType: string;
  recommendedAnswer: string;
  confidence: number;
  reasoningSummary: string;
  riskFlags: string[];
  rubricChecklist: string[];
  reviewLevel: WorkbenchAnalyzeResponseReviewLevel;
  humanReviewRequired: boolean;
  routingReceipt?: WorkbenchAnalyzeResponseRoutingReceipt;
}