# lib/api-zod/src/generated/types/workbenchRefuseCheckRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BridgeAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkbenchRefuseCheckRequest {
  requestText: string;
}