# lib/api-zod/src/generated/types/workbenchRefuseCheckResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BridgeAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkbenchRefuseCheckResponse {
  allowed: boolean;
  reason: string | null;
}